Make frame and prediction underlying pointer accessors abstract
Every backend will use a different data type. For example,
NCSDK may use 16 or 32 bit floats, others may use doubles, etc...
In order to be generic to multiple frameworks, the pointer
accessors will be type agnostic, and the framework should be
in sync with the app so that they match.